+++ unpacked/_next/static/chunks/pages/[[...default]].js
@@ -15,7 +15,7 @@
       var n = r(36112),
         u = r(10721),
         c = r(70079),
-        o = r(21129),
+        o = r(65689),
         f = r(35250);
       function i(e, t) {
         var r = Object.keys(e);
@@ -80,4 +80,4 @@
       (_N_E = e.O());
   },
 ]);
-//# sourceMappingURL=[[...default]]-d3a435e20d062ea2.js.map
+//# sourceMappingURL=[[...default]]-12836b0c1baa69d7.js.map